requirements.txt
praw
google-cloud-bigquery
python-dotenv
